third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/frame-src-blocked-expected.txt

CONSOLE ERROR: line 21: Refused to frame 'http://localhost:8000/security/resources/post-done.html' because it violates the following Content Security Policy directive: "frame-src 'self'".

This is a testharness.js-based test.
PASS The unredirected frame is blocked. 
Harness: the test ran to completion.
